requirements.txt
fastapi
paho-mqtt
pyjwt
passlib[argon2]
python-dotenv
sqlmodel
pydantic_settings
debugpy
mysql-connector-python
psycopg2-binary
uvicorn
dependency-injector